+++ tests/test_plugins_mark_as_read.py
@@ -48,10 +48,14 @@ def test_regex_mark_as_read_bad_metadata(make_reader, value):
 
     one = parser.feed(1, datetime(2010, 1, 1))
     parser.entry(1, 1, datetime(2010, 1, 1), title='match')
+    parser.entry(1, 2, datetime(2010, 1, 1), title='will be modified')
 
     reader.add_feed(one)
     reader.set_feed_metadata_item(one, '.reader.mark_as_read', value)
 
     reader.update_feeds()
 
-    assert [e.read for e in reader.get_entries()] == [False]
+    parser.entry(1, 2, datetime(2010, 1, 1), title='modified')
+    reader.update_feeds()
+
+    assert [e.read for e in reader.get_entries()] == [False, False]