+++ CHANGES.rst
@@ -6,6 +6,12 @@ Changelog
   :noindex:
 
 
+Version 1.21
+------------
+
+Unreleased
+
+
 Version 1.20
 ------------